feat: new settings side panel.
